+++ packages/design/src/constant/index.js
@@ -1,0 +1,1 @@
+export const DR_DESIGN_KEY = 'drDesign'