+++ packages/design/src/main.js
@@ -2,4 +2,8 @@ export { default as DrFormDesign } from './components/form'
 export { default as DrBasicDesign } from './components/basic'
 export { default as DrPageDesign } from './components/page'
 export { default as DrTableDesign } from './components/table'
-export { PLUGIN_ENUM, PreviewPlugin, ModulePlugin, DrawPlugin, ConfigurePlugin } from './plugins/plugin'
+export { PLUGIN_ENUM, PreviewPlugin, ModulePlugin, DrawPlugin, ConfigurePlugin, IconHandlePlugin } from './plugins/plugin'
+/**
+ * 设计器provide的key, 为插件提供全局数据支撑
+ */
+export { DR_DESIGN_KEY } from './constant'