+++ play/src/views/palyground/custom-form-design/plugins/icon/component.jsx
@@ -1,0 +1,10 @@
+import { ElIcon } from 'element-plus'
+import { Share } from '@element-plus/icons-vue'
+
+export default {
+  setup () {
+    return () => <ElIcon>
+      <Share></Share>
+    </ElIcon>
+  }
+}