+++ play/src/views/palyground/custom-form-design/plugins/icon/index.js
@@ -1,0 +1,9 @@
+import IconComponent from './component'
+import { IconHandlePlugin } from '@d-render/design'
+export class IconPlugin extends IconHandlePlugin {
+  constructor (options) {
+    super(options)
+    this.callback = options?.callback
+    this.Component = IconComponent
+  }
+}