+++ packages/d-render/CHANGELOG.md
@@ -1,5 +1,23 @@
 # d-render
 
+## 2.1.15-beta.12
+
+### Patch Changes
+
+- chore(d-render): 修改版本号后重新发布
+
+## 2.1.15-beta.11
+
+### Patch Changes
+
+- fix(cip-form): 修复`labelSuffix`属性不生效的问题
+
+## 2.1.15-beta.10
+
+### Patch Changes
+
+- polish(表单设计): 修改子表单样式
+
 ## 2.1.15-beta.9
 
 ### Patch Changes